refactor: apply hexagonal folder structure
