+++ src/main/java/io/pakland/mdas/githubstats/domain/model/Comment.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain;
+package io.pakland.mdas.githubstats.domain.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -14,30 +14,17 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 @Entity
-@Table(name = "repository")
-public class Repository {
+@Table(name = "comment")
+public class Comment {
 
   @Id
   @Column(name = "id", updatable = false, nullable = false)
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private int id;
 
-  @Column(name = "organization_id")
-  private int organizationId;
+  @Column(name = "review_by_user_id")
+  private int reviewByUserId;
 
-  @Column(name = "team_id")
-  private int teamId;
-
-  @Column(name = "name")
-  private String name;
-
-  @Column(name = "owner")
-  private String owner;
-
-  @Column(name = "merges_url")
-  private String mergesUrl;
-
-  @Column(name = "commits_url")
-  private String commitsUrl;
-
+  @Column(name = "length")
+  private int length;
 }
