+++ src/main/java/io/pakland/mdas/githubstats/domain/model/Repository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain;
+package io.pakland.mdas.githubstats.domain.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -14,14 +14,30 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 @Entity
-@Table(name = "pull_requests")
-public class PullRequests {
+@Table(name = "repository")
+public class Repository {
 
   @Id
   @Column(name = "id", updatable = false, nullable = false)
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private int id;
 
-  @Column(name = "repository_id")
-  private int repositoryId;
+  @Column(name = "organization_id")
+  private int organizationId;
+
+  @Column(name = "team_id")
+  private int teamId;
+
+  @Column(name = "name")
+  private String name;
+
+  @Column(name = "owner")
+  private String owner;
+
+  @Column(name = "merges_url")
+  private String mergesUrl;
+
+  @Column(name = "commits_url")
+  private String commitsUrl;
+
 }
