+++ src/main/java/io/pakland/mdas/githubstats/domain/ports/ICommit.java
@@ -1,6 +1,6 @@
-package io.pakland.mdas.githubstats.domain.repository;
+package io.pakland.mdas.githubstats.domain.ports;
 
-import io.pakland.mdas.githubstats.domain.Commit;
+import io.pakland.mdas.githubstats.domain.model.Commit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
