+++ src/main/java/io/pakland/mdas/githubstats/domain/ports/IRepository.java
@@ -1,0 +1,12 @@
+package io.pakland.mdas.githubstats.domain.ports;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+/**
+ * Add jdoc about the rep
+ */
+
+@Repository
+public interface IRepository extends JpaRepository<io.pakland.mdas.githubstats.domain.model.Repository,Integer> {
+}
