+++ src/main/java/io/pakland/mdas/githubstats/domain/ports/IReviewUser.java
@@ -1,6 +1,6 @@
-package io.pakland.mdas.githubstats.domain.repository;
+package io.pakland.mdas.githubstats.domain.ports;
 
-import io.pakland.mdas.githubstats.domain.ReviewUser;
+import io.pakland.mdas.githubstats.domain.model.ReviewUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
