--- src/main/java/io/pakland/mdas/githubstats/domain/repository/IRepository.java
@@ -1,12 +1,0 @@
-package io.pakland.mdas.githubstats.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * Add jdoc about the rep
- */
-
-@Repository
-public interface IRepository extends JpaRepository<io.pakland.mdas.githubstats.domain.Repository,Integer> {
-}
